refactor(agentic): make forecasting machinery domain-agnostic via DomainConfig
Promote the shared agentic engine so a new target series can be configured
without touching oil code, and drive the oil prompts/config from a single
`DomainConfig`. Pure refactor: no new features, no study driver.

Promoted to aieng/forecasting/methods/agentic/:
- strategy_state.py: generic `StrategyState` (+ Observation/Hypothesis/
  CalibrationCorrection/VersionEntry). Markdown title, default skill name, and
  frontmatter description are ClassVar params. Oil `WtiStrategyState` becomes a
  thin subclass pinning the WTI strings — committed SKILL.md renders identically.
- adaptive_skill_tools.py: `build_skill_tools(strategy_dir, state_type, ...)`
  generic over any StrategyState subclass. No import-time singletons. Oil
  skill_tools.py delegates and re-exports the sub-models; STORE/WTI_SKILL_TOOLS
  removed (unused outside docstrings).
- history.py: `compress_history` (was analyst_agent/agent.py); oil re-exports it.
- outputs.py: generic `ScenarioCard` / `ScenarioAgentForecastOutput` with the
  scenario-card numeric fields parameterized via `scenario_card_template_extra`.
  Oil `tasks.py` subclasses them, preserving `wti_range_60d`/`point_estimate_60d`
  and modality="discrete".
- domain.py: `DomainConfig` (identity, data/target, context-retrieval, strategy
  skill, skill dirs, vol-regime bands, tool bounds) plus render_* instruction
  templates and `build_analyst_config` / `build_adaptive_config`.

Oil implementation:
- energy_oil_forecasting/domain.py defines `OIL_DOMAIN`.
- analyst/adaptive/starter agent instructions now render from OIL_DOMAIN; the
  `build_wti_*` factories are thin delegating wrappers. Rendered oil prompts are
  byte-identical to the originals (verified against captured goldens).

Curriculum: `_vol_regime` vol bands extracted to a `bands` parameter with the
15/30/50 thresholds as `DEFAULT_VOL_REGIME_BANDS`.

Invariants held: oil agent.name / predictor_id strings unchanged; committed skill
artifacts render identically; nb05/nb06 import surfaces resolve. Regenerated the
concierge catalog/artifacts to index the new modules. __init__ re-exports the
promoted names; old oil import paths preserved.
