+++ implementations/getting_started/concierge_agent/context/artifacts/implementations__energy_oil_forecasting__specs__energy_oil_eval.yaml.md
@@ -5,8 +5,12 @@ kind: yaml
 ```yaml
 # Energy Oil Eval Spec — 2026 Prospective Competition
 #
-# Runs on 8 weekly origins from Feb 2, 2026 to Mar 23, 2026.
-# Covers the high-volatility Persian Gulf geopolitical price shock period.
+# Runs on 18 weekly origins from Feb 2, 2026 to Jun 1, 2026.
+# Covers the high-volatility Persian Gulf geopolitical price shock and its
+# aftermath. The end date is set to the latest origin whose longest horizon
+# (21 business days) still resolves against available data — keep it at
+# most 21 business days behind the most recent cached WTI price (see
+# scripts/fetch_wti.py) so every origin fully resolves.
 # Target is WTI Crude Oil price (yfinance ticker: CL=F).
 # Horizons: 5, 10, 21 business days.
 
@@ -14,8 +18,9 @@ spec_id: energy_oil_eval
 
 description: >-
   Prospective/out-of-sample evaluation period in 2026 for daily WTI crude oil.
-  Evaluates selected contender models on 8 weekly origins during the early 2026
-  geopolitical price shock to measure adaptive real-time forecasting performance.
+  Evaluates selected contender models on 18 weekly origins from the early 2026
+  geopolitical price shock through its aftermath, to measure adaptive
+  real-time forecasting performance.
 
 tasks:
   - task_id: wti_oil_price_forecast
@@ -27,7 +32,7 @@ tasks:
       projected 5, 10, and 21 trading days ahead.
 
 start: "2026-02-02"
-end: "2026-03-23"
+end: "2026-06-01"
 stride: 5
 warmup: 250
 ```
