+++ implementations/getting_started/concierge_agent/context/artifacts/implementations__energy_oil_forecasting__starter_agent____init__.py.md
@@ -5,10 +5,12 @@ kind: python
 ```python
 """WTI starter agent — a fresh, hackable template for your own exploration.
 
-Exports the toggle-driven :class:`AgentConfig` factory and the predictor
-convenience factory. See ``99_starter_agent.ipynb`` and ``agent.py``.
+Exports the toolbelt-driven :class:`AgentConfig` factory, the predictor
+convenience factory, and the :mod:`tools` module of per-tool factories you
+compose in the notebook. See ``99_starter_agent.ipynb`` and ``agent.py``.
 """
 
+from energy_oil_forecasting.starter_agent import tools
 from energy_oil_forecasting.starter_agent.agent import (
     build_starter_agent_config,
     build_starter_agent_predictor,
@@ -18,5 +20,6 @@ from energy_oil_forecasting.starter_agent.agent import (
 __all__ = [
     "build_starter_agent_config",
     "build_starter_agent_predictor",
+    "tools",
 ]
 ```
