+++ implementations/getting_started/concierge_agent/context/artifacts/implementations__energy_oil_forecasting__tasks.py.md
@@ -8,17 +8,20 @@ kind: python
 Implements the "one agent, three tasks" pattern: a single :class:`AgentConfig`
 identity with task-specific prompt builders and output schemas supplied via
 :class:`~aieng.forecasting.methods.agentic.predictor.AgentPredictor`.
+
+The scenario output classes are thin oil-branded subclasses of the shared,
+domain-agnostic :class:`~aieng.forecasting.methods.agentic.outputs.ScenarioCard`
+and :class:`~aieng.forecasting.methods.agentic.outputs.ScenarioAgentForecastOutput`.
+They pin the numeric scenario-card fields to the exact key names
+(``wti_range_60d``, ``point_estimate_60d``) the notebook caches key on.
 """
 
 from __future__ import annotations
 
 import json
-from datetime import datetime
 from typing import Any, ClassVar, Literal
 
-import pandas as pd
 from aieng.forecasting.data.context import ForecastContext
-from aieng.forecasting.evaluation.prediction import BinaryForecast, Prediction
 from aieng.forecasting.evaluation.task import ForecastingTask
 from aieng.forecasting.methods.agentic import (
     AgentPredictor,
@@ -26,7 +29,15 @@ from aieng.forecasting.methods.agentic import (
     DiscreteAgentForecastOutput,
 )
 from aieng.forecasting.methods.agentic.agent_factory import AgentConfig
-from aieng.forecasting.methods.agentic.outputs import AgentForecastOutput
+from aieng.forecasting.methods.agentic.outputs import (
+    AgentForecastOutput,
+)
+from aieng.forecasting.methods.agentic.outputs import (
+    ScenarioAgentForecastOutput as _BaseScenarioAgentForecastOutput,
+)
+from aieng.forecasting.methods.agentic.outputs import (
+    ScenarioCard as _BaseScenarioCard,
+)
 from aieng.forecasting.models import LITE_MODEL
 from energy_oil_forecasting.analyst_agent import (
     WtiPriceForecastPromptBuilder,
@@ -35,7 +46,7 @@ from energy_oil_forecasting.analyst_agent import (
     compress_history,
 )
 from energy_oil_forecasting.paths import SHOCK_HORIZON, SHOCK_THRESHOLD
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 
 # ── Task specification strings (embedded in user prompts for NB3) ───────────
@@ -72,89 +83,31 @@ class WtiMultitaskPromptBuilder(BaseModel):
         return json.dumps(payload, indent=2)
 
 
-class ScenarioCard(BaseModel):
-    """One scenario card from Task C agent output."""
+class ScenarioCard(_BaseScenarioCard):
+    """One scenario card from Task C agent output.
 
-    model_config = {"extra": "ignore"}
+    Adds the WTI-specific 60-day price range and point estimate to the generic
+    scenario card, under the exact field names the notebook caches expect.
+    """
 
-    name: str
-    description: str
-    probability: float = Field(ge=0.0, le=1.0)
     wti_range_60d: list[float]
     point_estimate_60d: float
-    key_drivers: list[str] = Field(default_factory=list)
 
 
-class ScenarioAgentForecastOutput(AgentForecastOutput):
-    """Track 2 scenario analysis output for the energy case study."""
+class ScenarioAgentForecastOutput(_BaseScenarioAgentForecastOutput):
+    """Track 2 scenario analysis output for the energy case study.
 
-    modality: ClassVar[Literal["continuous", "discrete"]] = "discrete"
-
-    model_config = {"extra": "ignore"}
+    Narrows the scenario-card type to :class:`ScenarioCard` and advertises the
+    WTI numeric fields in the prompt schema template.  Rendering and
+    :meth:`to_predictions` are inherited unchanged from the shared base.
+    """
 
     scenarios: list[ScenarioCard]
-    base_case: str
-    reasoning: str = ""
 
-    @classmethod
-    def prompt_schema_json(cls) -> str:
-        """Return a JSON template for use in agent instruction strings.
-
-        Returns
-        -------
-        str
-            Indented JSON string showing the exact structure the agent must
-            pass to ``set_model_response``.
-        """
-        template: dict[str, object] = {
-            "scenarios": [
-                {
-                    "name": "<string>",
-                    "description": "<string>",
-                    "probability": "<float in [0, 1]>",
-                    "wti_range_60d": ["<float_low>", "<float_high>"],
-                    "point_estimate_60d": "<float>",
-                    "key_drivers": ["<driver 1>", "<driver 2>"],
-                }
-            ],
-            "base_case": "<scenario name>",
-            "reasoning": "<paragraph>",
-        }
-        return json.dumps(template, indent=2)
-
-    def to_predictions(
-        self,
-        *,
-        task: ForecastingTask,
-        context: ForecastContext,
-        predictor_id: str,
-        metadata: dict[str, Any] | None = None,
-    ) -> list[Prediction]:
-        """Convert scenario output to a metadata-rich prediction (Track 2 display)."""
-        if len(task.horizons) != 1:
-            raise ValueError("Scenario agent output expects exactly one task horizon.")
-
-        horizon = task.horizons[0]
-        issued_at = datetime.utcnow()
-        offset = pd.tseries.frequencies.to_offset(task.frequency)
-        base_prob = float(sum(s.probability for s in self.scenarios))
-        prediction_metadata: dict[str, Any] = dict(metadata) if metadata is not None else {}
-        prediction_metadata["scenarios"] = [s.model_dump() for s in self.scenarios]
-        prediction_metadata["base_case"] = self.base_case
-        if self.reasoning.strip():
-            prediction_metadata["rationale"] = self.reasoning
-
-        return [
-            Prediction(
-                predictor_id=predictor_id,
-                task_id=task.task_id,
-                issued_at=issued_at,
-                as_of=context.as_of,
-                forecast_date=(pd.Timestamp(context.as_of) + offset * horizon).to_pydatetime(),
-                payload=BinaryForecast(probability=min(base_prob, 1.0)),
-                metadata=prediction_metadata,
-            )
-        ]
+    scenario_card_template_extra: ClassVar[dict[str, object]] = {
+        "wti_range_60d": ["<float_low>", "<float_high>"],
+        "point_estimate_60d": "<float>",
+    }
 
 
 # Task specification strings embedded in user prompts for NB3.
