+++ implementations/getting_started/concierge_agent/skills/repo-navigation/references/catalog-summary.yaml
@@ -1,20 +1,20 @@
 # Concierge catalog summary (regenerated by scripts/build_concierge_context.py)
 source_url: https://github.com/VectorInstitute/agentic-forecasting
 branch: main
-built_at: '2026-06-30T15:09:29+00:00'
-git_ref: 0ac6b3098bdb529c08f2445895d82490de2404fd
-entry_count: 196
+built_at: '2026-07-16T00:57:38+00:00'
+git_ref: 9715e2eeb6479c83eb7fd35acc97ebac26db8988
+entry_count: 202
 domains:
   docs: 4
   core.root: 3
   core.data: 12
   core.documents: 5
   core.evaluation: 9
-  core.methods: 26
+  core.methods: 30
   impl.README.md: 1
   impl.__init__.py: 1
   impl.boc_rate_decisions: 27
-  impl.energy_oil_forecasting: 45
+  impl.energy_oil_forecasting: 47
   impl.food_price_forecasting: 21
   impl.getting_started: 16
   impl.sp500_forecasting: 19
